moved the locks.py file shoonya_backend
